+++ NEWS.md
@@ -1,5 +1,10 @@
 # GoTTY releases
 
+## v1.3.0 (Unreleased)
+
+ * Links in the tty are now clickable.
+ * Use WebGL for rendering by default.
+
 ## v1.2.0
 
  * Pass BUILD\_OPTIONS to gox, too, so release artifacts have version info included.